[level 2] Title: 멀리 뛰기, Time: 0.06 ms, Memory: 74.3 MB -BaekjoonHub

+++ 프로그래머스/lv2/12914. 멀리 뛰기/멀리 뛰기.java
@@ -1,6 +1,6 @@
 class Solution {
-    public long solution(int n) {
-        long[] dp = new long[n];
+    public int solution(int n) {
+        int[] dp = new int[n];
         dp[0]=1;
         if(n>1) dp[1] = 2;
         for(int i=2; i<n; i++)
